requirements.txt
cycler==0.10.0
ImageHash==4.0
kiwisolver==1.0.1
matplotlib==3.0.2
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
pyparsing==2.3.1
python-dateutil==2.7.5
PyWavelets==1.0.1
scipy==1.2.0
six==1.12.0
tqdm==4.29.1
